# contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Travel Tour Booking API
  version: 0.16.0
  description: |
    API cho website bán tour du lịch Việt Nam, hai thị trường (DK, VN),
    hai ngôn ngữ (da, vi).
servers:
  - url: /
    description: Cùng gốc với ứng dụng
...
paths:
...
  /api/v1/admin/products/{id}/itinerary:
    get:
      tags: [admin]
      operationId: getAdminItinerary
      summary: Lịch trình từng ngày, kèm **mọi** bản dịch
      description: |
        Khác `GET /api/v1/{market}/products/{slug}/itinerary` ở đúng chỗ mọi bề
        mặt quản trị khác: trả **tất cả** locale cạnh nhau, không trả một bản
        theo `Accept-Language`. Người dịch cần nhìn hai bản một lúc.

        `404` với `COMBO` và `DAY_TOUR` — hai loại đó không có lịch trình theo
        ngày, cùng lý do và cùng mã với bề mặt công khai.
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      responses:
        '200':
          description: Lịch trình, sắp theo `dayNumber`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminItinerary' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

    put:
      tags: [admin]
      operationId: saveItinerary
      summary: Thay **toàn bộ** lịch trình, kèm bản ngôn ngữ nguồn
      description: |
        `PUT` thay sạch, không `POST` từng ngày. Lịch trình là **một mảng có
        ràng buộc giữa các phần tử** — số ngày phải bằng `durationDays`, và
        `dayNumber` phải liền mạch từ 1. Luật đó chỉ kiểm được khi nhìn cả mảng,
        nên nhận cả mảng. Cùng khuôn với `price-tiers`, `prices` và `roles`.

        Ngày biến mất khỏi mảng bị **xoá mềm**, không xoá cứng: index duy nhất
        `ux_itinerary_day` là index bộ phận đúng để việc bỏ một ngày rồi thêm
        lại ngày khác cùng số là thao tác biên tập bình thường (`docs/12` mục
        2.2). Ngày còn trong mảng thì giữ nguyên `id` — bản dịch treo vào id đó.

        Thân yêu cầu mang luôn **bản ngôn ngữ nguồn** (`da`, ADR-004), cùng lý
        do với `POST /admin/products`: một ngày lịch trình không có tiêu đề thì
        không phải một ngày lịch trình. Bản `vi` đặt sau bằng endpoint bên dưới.

        Ba luật liên trường, cả ba trả `400 VALIDATION_FAILED` kèm `fields` chỉ
        đúng ô: số ngày lệch `durationDays`, `dayNumber` trùng hoặc thủng, và
        `destinationId`/`hotelId` trỏ vào thứ không tồn tại.
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminItinerarySave' }
      responses:
        '200':
          description: Lịch trình sau khi lưu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminItinerary' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/products/{id}/itinerary/translations/{locale}:
    put:
      tags: [admin]
      operationId: saveItineraryTranslation
      summary: Dịch lịch trình sang một ngôn ngữ
      description: |
        Chỉ đụng tới chữ. **Không** tạo, không xoá, không đổi thứ tự ngày —
        cấu trúc là việc của `PUT /itinerary`, và người dịch không được đổi nó.

        `dayNumber` gửi lên phải khớp **đúng** tập ngày đang có. Thiếu một ngày
        hay thừa một ngày đều là `400`: dịch nửa chừng rồi lưu sẽ để lại một
        lịch trình mà locale này thấy thủng ngày, còn locale kia thì không.

        Quyền phụ thuộc locale, cùng ma trận với bản dịch sản phẩm (`docs/22`
        mục 2.1): `EDITOR` ghi bản `da`, `TRANSLATOR` ghi bản `vi`, `ADMIN` ghi
        cả hai.
      parameters:
        - $ref: '#/components/parameters/AdminId'
        - name: locale
          in: path
          required: true
          schema:
            type: string
            enum: [da, vi]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminItineraryTranslationSave' }
      responses:
        '200':
          description: Lịch trình sau khi lưu bản dịch
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminItinerary' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /api/v1/admin/destinations/{id}/media:
    get:
      tags: [admin]
      operationId: getDestinationMedia
      summary: Ảnh và video của một điểm đến, theo thứ tự
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      responses:
        '200':
          description: Danh sách theo `sort_order`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMediaList' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

    put:
      tags: [admin]
      operationId: saveDestinationMedia
      summary: Thay **toàn bộ** danh sách ảnh và video của điểm đến
      description: |
        `PUT` thay sạch, cùng khuôn với `price-tiers`, `tags` và `roles`: thứ tự
        là một thuộc tính của **cả danh sách**, nên sửa từng phần tử không diễn
        đạt được "đổi chỗ hai tấm".

        Gỡ một tệp khỏi đây **không xoá tệp**, chỉ xoá dòng nối — chứng từ giấy
        phép của nó ở lại (`docs/12` mục 4.6). Tệp cùng lúc dùng ở nhiều nơi là
        chuyện bình thường.

        `assetId` trỏ vào tệp không tồn tại, hoặc trùng nhau trong cùng một mảng,
        đều là `400 VALIDATION_FAILED` chỉ đúng phần tử sai.
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaOrder' }
      responses:
        '200':
          description: Danh sách sau khi lưu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMediaList' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/products/{id}/images:
    get:
      tags: [admin]
      operationId: getProductImages
      summary: Bộ ảnh của một sản phẩm, theo thứ tự
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      responses:
        '200':
          description: Danh sách theo `sort_order`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMediaList' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

    put:
      tags: [admin]
      operationId: saveProductImages
      summary: Thay **toàn bộ** bộ ảnh của sản phẩm
      description: |
        Cùng khuôn với `media` của điểm đến. Khác một luật: bộ ảnh sản phẩm
        **chỉ nhận `IMAGE`** — bảng nối là `product_image` và nó vẫn đúng tên.
        Gửi một `VIDEO` vào đây là `400` chỉ đúng phần tử đó.

        Thứ tự ở đây là thứ tự khách thấy trong bộ ảnh ở trang chi tiết. Ảnh
        **không có `alt` ở locale nào thì biến mất khỏi locale đó** — nó vẫn nằm
        trong danh sách này, chỉ là không tới được mắt khách đọc thứ tiếng ấy.
      parameters: [{ $ref: '#/components/parameters/AdminId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaOrder' }
      responses:
        '200':
          description: Bộ ảnh sau khi lưu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMediaList' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/media/upload-url:
    post:
      tags: [admin]
      operationId: createMediaUploadUrl
      summary: Xin URL đã ký để tải một tệp lên kho ảnh
      description: |
        **Backend không nhận byte.** Nó ký một URL có hạn, trình duyệt `PUT`
        thẳng tệp lên kho, rồi gọi `POST /api/v1/admin/media` để ghi bản ghi —
        ADR-011 mục 4. Đẩy tệp qua backend nghĩa là một tour 40 ảnh đi qua bộ nhớ
        của API hai lần, và tiến trình Java là chỗ đắt nhất để làm việc đó.

        **`path` do máy chủ đặt, không nhận từ client.** Tên tệp khách gửi lên
        chỉ dùng để lấy phần đuôi; đường dẫn thật là `<thư mục>/<uuid>.<đuôi>`.
        Cho client chọn đường dẫn là cho nó ghi đè ảnh của bản ghi khác.

        URL ký bằng **địa chỉ công khai** của kho chứ không phải địa chỉ nội bộ:
        chữ ký SigV4 bao gồm cả host, nên ký sai host thì trình duyệt nhận `403`
        mà không nói vì sao (`docs/15` mục 2).

        Ký xong mà không ai tải lên thì **không có gì được tạo** — không dòng
        `media_asset` nào, không tệp nào. Tệp tải lên mà không gọi bước hai là
        tệp mồ côi, dọn theo ADR-011 mục 6 việc 6.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaUploadUrlRequest' }
      responses:
        '200':
          description: URL đã ký, dùng được tới `expiresAt`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaUploadUrl' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/media:
    post:
      tags: [admin]
      operationId: createMedia
      summary: Ghi bản ghi ảnh sau khi tệp đã lên kho
      description: |
        Bước hai của luồng tải lên. Gọi sau khi `PUT` vào `uploadUrl` trả về
        `2xx`.

        **Chứng từ giấy phép là bắt buộc với ảnh không tự chụp.** `source` khác
        `SELF` mà thiếu `licenceRef` thì `400 VALIDATION_FAILED` chỉ đúng ô
        `licenceRef` — ràng buộc `ck_media_licence` của CSDL cũng nói y như vậy,
        và ảnh tạm là thứ ở lại lâu nhất (`docs/24` mục 8).

        **`alt` nhập bằng ngôn ngữ nguồn** (`da`, ADR-004). Bản `vi` đặt sau
        bằng `PUT /api/v1/admin/media/{id}/translations/{locale}`, và tới lúc đó
        ảnh mới hiện ở trang tiếng Việt: ảnh thiếu `alt` ở locale nào thì **biến
        mất** ở locale đó, đúng luật không fallback cho nội dung bán hàng.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminMediaCreate' }
      responses:
        '201':
          description: Ảnh vừa ghi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMedia' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/media/{id}/translations/{locale}:
    put:
      tags: [admin]
      operationId: saveMediaTranslation
      summary: Đặt chữ thay ảnh cho một ngôn ngữ
      description: |
        Quyền phụ thuộc **locale đang sửa**, cùng ma trận với bản dịch sản phẩm
        (`docs/22` mục 2.1): `EDITOR` ghi bản `da`, `TRANSLATOR` ghi bản `vi`,
        `ADMIN` ghi cả hai.

        `alt` là **nội dung phải dịch**, không phải siêu dữ liệu kỹ thuật — nó là
        thứ người khiếm thị đọc thay cho ảnh (`docs/24` mục 6).
      parameters:
        - $ref: '#/components/parameters/AdminId'
        - name: locale
          in: path
          required: true
          schema:
            type: string
            enum: [da, vi]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminMediaTranslationInput' }
      responses:
        '200':
          description: Ảnh sau khi đặt chữ thay
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMedia' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    AdminId:
      name: id
      in: path
      required: true
      description: |
        Định danh nội bộ. Bề mặt quản trị dùng `id`, bề mặt công khai dùng `slug`
        — `docs/13` mục 4.
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Lỗi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    AdminItinerary:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items: { $ref: '#/components/schemas/AdminItineraryDay' }
          description: |
            Sắp theo `dayNumber`. **Có thể rỗng** — sản phẩm chưa nhập lịch trình
            là trạng thái hợp lệ, và màn hình phải dựng được khi rỗng.
    AdminItineraryDay:
      type: object
      required: [dayNumber, translations]
      properties:
        dayNumber: { type: integer, format: int32, minimum: 1 }
        destinationId:
          type: string
          format: uuid
          description: |
            Nơi **ngủ đêm**. Vắng mặt với ngày bay hoặc ngày trên tàu — đó là
            trạng thái hợp lệ, không phải dữ liệu thiếu.
        destinationName:
          type: string
          description: Tên ở **ngôn ngữ nguồn**, trả sẵn để bảng đọc được mà không phải tra thêm.
        hotelId: { type: string, format: uuid }
        hotelName:
          type: string
          description: Tên riêng, **không dịch** — `docs/24` mục 5.
        translations:
          type: array
          items: { $ref: '#/components/schemas/AdminItineraryDayText' }
          description: |
            Mọi locale đang có chữ cho ngày này. Locale vắng mặt nghĩa là ngày
            này **chưa dịch** sang đó, và cả lịch trình sẽ không hiện ở locale
            ấy — bề mặt công khai `INNER JOIN` bảng dịch.
    AdminItineraryDayText:
      type: object
      required: [locale, title, description, isSource]
      properties:
        locale: { type: string }
        title: { type: string }
        description: { type: string }
        isSource:
          type: boolean
          description: Bản này có phải ngôn ngữ nguồn không — ADR-004.
    AdminItinerarySave:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items: { $ref: '#/components/schemas/AdminItineraryDayInput' }
          description: |
            **Toàn bộ** lịch trình. Ngày không có trong mảng này bị xoá mềm.
    AdminItineraryDayInput:
      type: object
      required: [dayNumber, title, description]
      properties:
        dayNumber: { type: integer, format: int32, minimum: 1, maximum: 60 }
        destinationId: { type: string, format: uuid }
        hotelId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, minLength: 1 }
    AdminItineraryTranslationSave:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items: { $ref: '#/components/schemas/AdminItineraryDayTranslationInput' }
    AdminItineraryDayTranslationInput:
      type: object
      required: [dayNumber, title, description]
      properties:
        dayNumber: { type: integer, format: int32, minimum: 1, maximum: 60 }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, minLength: 1 }
    MediaKind:
      type: string
      description: |
        Loại tệp trong kho. Ảnh bìa của một video cũng là một bản ghi `IMAGE`
        riêng — nó có giấy phép và `alt` của chính nó.
      enum: [IMAGE, VIDEO]
...
    MediaOrder:
      type: object
      description: |
        Danh sách tệp theo **thứ tự**, thay toàn bộ. Thứ tự trong mảng chính là
        `sort_order`; không có trường thứ tự riêng để không ai phải giữ hai nguồn
        sự thật cho cùng một việc.
      required: [assetIds]
      properties:
        assetIds:
          type: array
          items: { type: string, format: uuid }
          maxItems: 40
    AdminMediaList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminMedia' }
    MediaSource:
      type: string
      description: |
        Ảnh từ đâu ra. Quyết định chứng từ giấy phép có bắt buộc hay không —
        `docs/24` mục 8, ràng buộc `ck_media_licence`.
      enum: [SELF, PARTNER, PURCHASED, CUSTOMER]
    MediaUploadUrlRequest:
      type: object
      required: [fileName, contentType, byteSize]
      properties:
        fileName:
          type: string
          maxLength: 260
          description: |
            Chỉ dùng để lấy **phần đuôi**. Đường dẫn thật do máy chủ đặt.
        contentType:
          type: string
          description: |
            Danh mục đóng, và cố ý hẹp. Kho ảnh mở quyền đọc ẩn danh để
            `next/image` tải được, nên mọi thứ tải lên đó là **tệp công khai** —
            danh sách này là thứ ngăn nó thành nơi chứa tệp bất kỳ.
          enum: [image/jpeg, image/png, image/webp, image/avif, video/mp4, video/webm]
        byteSize:
          type: integer
          format: int64
          minimum: 1
          maximum: 209715200
          description: |
            Trần ở đây là trần **chung** (200 MB). Trần thật phụ thuộc loại tệp và
            máy chủ kiểm riêng: **10 MB cho ảnh, 200 MB cho video**. Vượt trần thì
            `400 VALIDATION_FAILED` chỉ đúng ô `byteSize`, kèm `max` là trần của
            đúng loại đang tải.

            Hai trần chứ không một, vì ký một URL là hứa nhận từng ấy byte vào
            đĩa VPS và không có ai đứng giữa để dừng lại giữa chừng — mà một ảnh
            10 MB thì đã là ảnh chưa ai nén, còn một video 10 MB thì chỉ dài vài
            giây.
        folder:
          type: string
          enum: [tour, diem-den, khach-san, bai-viet]
          default: tour
          description: Thư mục đầu trong kho. Chỉ để tệp dễ tìm bằng mắt.
    MediaUploadUrl:
      type: object
      required: [uploadUrl, path, expiresAt]
      properties:
        uploadUrl:
          type: string
          description: |
            `PUT` thẳng tệp vào đây, kèm header `Content-Type` **đúng bằng**
            `contentType` đã xin. Lệch một chữ là chữ ký không khớp và kho trả
            `403`.
        path:
          type: string
          description: |
            Đường dẫn **tương đối** trong bucket. Gửi lại nguyên văn ở bước hai;
            đây cũng chính là thứ lưu vào `media_asset.path` (ADR-011 mục 2).
        expiresAt: { type: string, format: date-time }
    AdminMediaCreate:
      type: object
      required: [path, width, height, byteSize, source, alt]
      properties:
        kind:
          allOf: [{ $ref: '#/components/schemas/MediaKind' }]
          default: IMAGE
        path: { type: string, maxLength: 500 }
        contentType:
          type: string
          maxLength: 64
          description: |
            **Bắt buộc với `VIDEO`** — thẻ `<video>` cần nó để chọn bộ giải mã.
            Ảnh thì bỏ trống được: trình duyệt tự đoán từ byte đầu.
        durationSeconds:
          type: integer
          format: int32
          minimum: 1
          description: |
            **Bắt buộc với `VIDEO`**, và phải rỗng với `IMAGE` — ràng buộc
            `ck_media_video` cưỡng chế cả hai chiều.
        posterAssetId:
          type: string
          format: uuid
          description: |
            Ảnh bìa, **bắt buộc với `VIDEO`**. Phải là một `media_asset` có
            `kind = IMAGE`; trỏ vào một video khác là `400`.

            Vì sao bắt buộc: thiếu nó thì trình duyệt tải những megabyte đầu của
            video chỉ để vẽ một khung hình tĩnh, và trên VPS không CDN đó là băng
            thông trả cho thứ khách chưa bấm xem (`docs/24` mục 7).
        width: { type: integer, format: int32, minimum: 1 }
        height: { type: integer, format: int32, minimum: 1 }
        byteSize: { type: integer, format: int64, minimum: 1 }
        source: { $ref: '#/components/schemas/MediaSource' }
        licenceRef:
          type: string
          maxLength: 200
          description: |
            **Bắt buộc khi `source` khác `SELF`.** Số hợp đồng, đường dẫn tới thư
            cho phép, hoặc mã đơn mua — thứ tra lại được sau hai năm.
        licenceScope: { type: string, maxLength: 200 }
        licenceUntil: { type: string, format: date }
        personConsent:
          type: boolean
          default: false
          description: Có người nhận diện được trong ảnh và đã đồng ý — `docs/31`.
        alt:
          type: string
          minLength: 1
          maxLength: 300
          description: Chữ thay ảnh, bằng **ngôn ngữ nguồn** `da`.
    AdminMediaTranslationInput:
      type: object
      required: [alt]
      properties:
        alt: { type: string, minLength: 1, maxLength: 300 }
    AdminMedia:
      type: object
      required: [id, kind, path, url, width, height, byteSize, source, translatedLocales]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/MediaKind' }
        contentType: { type: string }
        durationSeconds: { type: integer, format: int32 }
        posterUrl:
          type: string
          description: Địa chỉ công khai của ảnh bìa. Chỉ có ở `VIDEO`.
        path:
          type: string
          description: Đường dẫn tương đối, đúng như lưu trong CSDL.
        url:
          type: string
          description: |
            Địa chỉ công khai đầy đủ, ghép từ `path` và cấu hình kho. Trả sẵn để
            màn hình quản trị hiện ảnh xem trước mà không phải tự ghép — ghép ở
            frontend là chép một quy tắc của backend sang chỗ thứ hai.
        width: { type: integer, format: int32 }
        height: { type: integer, format: int32 }
        byteSize: { type: integer, format: int64 }
        source: { $ref: '#/components/schemas/MediaSource' }
        licenceRef: { type: string }
        translatedLocales:
          type: array
          items: { type: string }
          description: |
            Locale đã có `alt`. Locale **không** nằm ở đây thì ảnh này vô hình
            trên bản ngôn ngữ đó — màn hình quản trị hiện cảnh báo dựa vào chính
            trường này.
    ErrorCode:
      type: string
      description: |
        **Danh mục đóng.** Mọi mã lỗi hệ thống này phát ra đều nằm ở đây, và đây
        là nguồn sự thật cho cả hai phía: backend không viết chuỗi mã ra tay,
        frontend phải dịch đủ danh sách này — thiếu một mã là lỗi biên dịch chứ
        không phải một dòng mã lạ hiện ra trước mặt người dùng.

        Bảng đối chiếu mã ↔ HTTP ↔ khi nào xảy ra: `docs/13` mục 5.1.

        Thêm mã mới là **đổi hợp đồng**: sửa ở đây, sinh lại, rồi cả hai phía
        cùng đỏ cho tới khi được xử lý. Đó là chủ ý.
      example: NOT_FOUND
      enum:
        # 400 — dữ liệu vào
        - VALIDATION_FAILED
        - UNSUPPORTED_LOCALE
        - PRODUCT_TYPE_BLOCK_MISMATCH
        - DURATION_DAYS_RULE_VIOLATED
        - CABIN_CATEGORY_NOT_ALLOWED
        - UNKNOWN_PAX_TYPE
        - PRICE_TIER_NOT_CONTIGUOUS
        # 401 · 403 · 404
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        # 409 — hợp lệ, nhưng trạng thái hiện tại không cho
        - DEPARTURE_SOLD_OUT
        - SEAT_HOLD_EXPIRED
        - DEPARTURE_CLOSED
        - QUOTE_EXPIRED
        - QUOTE_NOT_ACCEPTABLE
        - CAPACITY_BELOW_BOOKED
        - BOOKING_TRANSITION_NOT_ALLOWED
        - PRODUCT_HAS_ACTIVE_BOOKINGS
        - SINGLE_PRICE_MISSING
        - DESTINATION_IN_USE
        - LAST_ADMIN
        - IDEMPOTENCY_KEY_REUSED
        # 422 — yêu cầu này không bao giờ hợp lệ
        - LEAD_TIME_NOT_MET
        - PRODUCT_NOT_QUOTABLE
        - PRODUCT_NOT_BOOKABLE
        - PARTY_SIZE_OUT_OF_RANGE
        # 500
        - INTERNAL_ERROR
...
    ErrorResponse:
      type: object
      description: Mã lỗi kèm tham số. Frontend dịch — backend không trả câu tiếng người.
      required: [code]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        params:
          type: object
          additionalProperties: true
        fields:
          type: array
          description: |
            Chỉ có ở `VALIDATION_FAILED`, và chỉ khi lỗi định vị được tới từng
            trường (`docs/13` mục 5). Luật liên trường không nằm ở đây — chúng có
            mã lỗi riêng, ví dụ `DURATION_DAYS_RULE_VIOLATED`.

            Rỗng thì **bỏ hẳn khỏi JSON**, không trả `[]` — quy tắc trường rỗng
            của `docs/13` mục 4.
          items:
            $ref: '#/components/schemas/FieldError'
        traceId:
          type: string
...